Get artists(from db) done.
