+++ daemon/enumtype/datasourcetype.py
@@ -1,0 +1,10 @@
+from enum import Enum, unique
+
+
+@unique
+class DataSourceType(Enum):
+    Spotify = "SPOTIFY"
+    DataBase = "DB"
+
+    def describe(self):
+        return self.name, self.value
